pyproject.toml
[tool.poetry]
name = "crm-api"
version = "0.1.0"
description = ""
authors = ["migue <miguel.lau@etecsa.cu>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
alembic = "^1.8.1"
psycopg2 = "^2.9.3"
SQLAlchemy = "^1.4.41"
PyJWT = "^2.5.0"
python-dotenv = "^0.21.0"
FastAPI-SQLAlchemy = "^0.2.1"
passlib = "^1.7.4"
itsdangerous = "^2.1.2"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
